WebKit/LayoutTests/http/tests/performance-timing/paint-timing/first-paint-only.html

<!DOCTYPE html>
<head>
<title>Performance Paint Timing Test: FP only</title>
</head>
<body>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <div id="main"></div>

    <script>
    async_test(function (t) {
        const div = document.createElement("div");
        div.style.width = "100px";
        div.style.height = "100px";
        div.style.backgroundColor = "red";
        div.style.color = "blue";
        document.getElementById("main").appendChild(div);

        testRunner.capturePixelsAsyncThen(t.step_func_done(function() {
            const bufferedEntries = performance.getEntriesByType('paint');
            assert_equals(bufferedEntries.length, 1, "FP only.");
            assert_equals(bufferedEntries[0].entryType, "paint");
            assert_equals(bufferedEntries[0].name, "first-paint");
        }));
    }, "Performance first paint timing entry exists. No first contentful paint.");
    </script>
</body>
</html>
